the TS_CMD_VSS_RATIO* id values overflow signed 16 bit int on avr
switch to unsigned 16 bit value

define TS_CMD_VSS_60KMH  39168 //0x99x00
define TS_CMD_VSS_RATIO1 39169
define TS_CMD_VSS_RATIO2 39170
define TS_CMD_VSS_RATIO3 39171
define TS_CMD_VSS_RATIO4 39172
define TS_CMD_VSS_RATIO5 39173
define TS_CMD_VSS_RATIO6 39174

+++ speeduino/TS_CommandButtonHandler.ino
@@ -19,7 +19,7 @@
  * 
  * @param buttonCommand The command number of the button that was clicked. See TS_CommendButtonHandler.h for a list of button IDs
  */
-void TS_CommandButtonsHandler(int buttonCommand)
+void TS_CommandButtonsHandler(uint16_t buttonCommand)
 {
   switch (buttonCommand)
   {
